requirements.txt
pandas
nltk
spacy
pypdf
python-docx
streamlit